requirements.txt
streamlit==1.28.1
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.1
google-cloud-functions==1.13.4
pypdf2==3.0.1
python-docx==1.1.0
google-generativeai==0.3.2
pandas==2.1.3
python-dotenv==1.0.0
gunicorn==21.2.0
